Changed firmware reporting format to avoid JSON message length going over 1024 bytes on SMP nodes. Changed solar and battery voltage methods to round values to 2 decimal places.
Removed BT code, added an MQTT login if no messages were uplinked so config changes are still received.

Added SIM CCID to source_ids object.

+++ firmware/wombat/include/power_monitoring/battery.h
@@ -29,7 +29,7 @@ private:
 public:
     static void begin();
 
-    static float get_voltage();
+    static double get_voltage();
     static float get_current();
 
     static void sleep();
